Refactor FindResult converter: use attribute, extract shared rewriter, fix tests
- Register FindResultConverterFactory via [JsonConverter] attribute on
  FindResult<T> (matching ReadItemResponse pattern) instead of mutating
  JsonSerializerOptions in CouchClient, avoiding unnecessary copy of
  the shared JsonSerializerOptions.Web singleton.
- Extract _id/_rev rewriting logic into DocumentRewriter shared by both
  FindResultConverter and ReadItemResponseConverter.
- Default bookmark to string.Empty when missing instead of null-forgiving.
- Add Rev property to test Rebel model and assert _rev mapping in tests.
- Revert unrelated README example change.
- Remove unused jsonString variable and GetPropertyName from
  ReadItemResponseConverter.
