+++ src/CouchDB.Driver/Converters/FindResultConverter.cs
@@ -1,4 +1,3 @@
-using System.Buffers;
 using System.Text.Json;
 using CouchDB.Driver.DTOs;
 using CouchDB.Driver.Types;
@@ -37,7 +36,7 @@ internal class FindResultConverter<TItem> : JsonConverter<FindResult<TItem>>
         }
 
         TItem[]? docs = null;
-        string bookmark = null!;
+        string? bookmark = null;
         ExecutionStats? executionStats = null;
         string? warning = null;
 
@@ -62,7 +61,7 @@ internal class FindResultConverter<TItem> : JsonConverter<FindResult<TItem>>
                     docs = ReadDocs(ref reader, options);
                     break;
                 case "bookmark":
-                    bookmark = reader.GetString()!;
+                    bookmark = reader.GetString();
                     break;
                 case "execution_stats":
                     executionStats = JsonSerializer.Deserialize<ExecutionStats>(ref reader, options);
@@ -76,7 +75,7 @@ internal class FindResultConverter<TItem> : JsonConverter<FindResult<TItem>>
             }
         }
 
-        return new FindResult<TItem>(docs ?? [], bookmark!, executionStats, warning);
+        return new FindResult<TItem>(docs ?? [], bookmark ?? string.Empty, executionStats, warning);
     }
 
     public override void Write(Utf8JsonWriter writer, FindResult<TItem> value, JsonSerializerOptions options)
@@ -99,77 +98,9 @@ internal class FindResultConverter<TItem> : JsonConverter<FindResult<TItem>>
                 break;
             }
 
-            items.Add(ReadDocument(ref reader, options));
+            items.Add(DocumentRewriter.RewriteDocument<TItem>(ref reader, options));
         }
 
         return items.ToArray();
-    }
-
-    private static TItem ReadDocument(ref Utf8JsonReader reader, JsonSerializerOptions options)
-    {
-        if (reader.TokenType != JsonTokenType.StartObject)
-        {
-            throw new JsonException("Expected start of object");
-        }
-
-        var bufferWriter = new ArrayBufferWriter<byte>();
-        using var writer = new Utf8JsonWriter(bufferWriter);
-
-        writer.WriteStartObject();
-
-        while (reader.Read())
-        {
-            if (reader.TokenType == JsonTokenType.EndObject)
-            {
-                break;
-            }
-
-            if (reader.TokenType != JsonTokenType.PropertyName)
-            {
-                throw new JsonException("Expected property name");
-            }
-
-            var propertyName = reader.GetString()!;
-            reader.Read();
-
-            switch (propertyName)
-            {
-                case "_id":
-                    var id = reader.GetString()!;
-                    WriteTransformedProperty(writer, "id", id, options);
-                    break;
-                case "_rev":
-                    var rev = reader.GetString()!;
-                    WriteTransformedProperty(writer, "rev", rev, options);
-                    break;
-                default:
-                    writer.WritePropertyName(propertyName);
-                    JsonSerializer.Serialize(writer, JsonSerializer.Deserialize<JsonElement>(ref reader), options);
-                    break;
-            }
-        }
-
-        writer.WriteEndObject();
-        writer.Flush();
-
-        var jsonReader = new Utf8JsonReader(bufferWriter.WrittenSpan);
-        return JsonSerializer.Deserialize<TItem>(ref jsonReader, options)!;
-    }
-
-    private static void WriteTransformedProperty(Utf8JsonWriter writer, string name, string value,
-        JsonSerializerOptions options)
-    {
-        if (options.PropertyNamingPolicy == null)
-        {
-            writer.WritePropertyName(name);
-            writer.WriteStringValue(value);
-            writer.WritePropertyName(char.ToUpper(name[0]) + name[1..]);
-            writer.WriteStringValue(value);
-        }
-        else
-        {
-            writer.WritePropertyName(options.PropertyNamingPolicy.ConvertName(name));
-            writer.WriteStringValue(value);
-        }
     }
 }
